requirements.txt
requests~=2.32.3
beautifulsoup4~=4.12.3
selenium~=4.22.0
python-dotenv~=1.0.1
pytz~=2024.1
fastapi~=0.111.0
scikit-learn~=1.4.2
uvicorn~=0.30.1
starlette~=0.37.2
redis-om==0.3.1
pandas~=2.2.1
konlpy~=0.6.0
numpy~=1.26.4
pydantic~=2.7.4
pyee~=11.1.0
pytest~=8.2.2